Combine to a single project and make some minor changes
