+++ src/PerformanceTests/Tests/Enabled/CSharp_CA1831.cs
@@ -4,18 +4,20 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 using BenchmarkDotNet.Attributes;
 using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.NetCore.CSharp.Analyzers.Runtime;
+using Microsoft.NetCore.Analyzers.Performance;
 using PerformanceTests.Utilities;
 using PerfUtilities;
 
 namespace CSharpPerformanceTests.Enabled
 {
-    public class CSharp_CA2014
+    public class CSharp_CA1831
     {
         [IterationSetup]
-        public static void CreateEnvironmentCA2014()
+        public static void CreateEnvironmentCA1831()
         {
             var sources = new List<(string name, string content)>();
             for (var i = 0; i < Constants.Number_Of_Code_Files; i++)
@@ -23,27 +25,14 @@ namespace CSharpPerformanceTests.Enabled
                 var name = "TypeName" + i;
                 sources.Add((name, @$"
 using System;
-unsafe class {name} {{
-    private static void ForLoop() {{
-        for (int i = 0; i < 10; i++)
-        {{
-            Span<byte> span = stackalloc byte[1024]|;
-        }}
-    }}
 
-    private static void WhileLoop() {{
-        while (true)
-        {{
-            Span<char> span = stackalloc char[1024];
-        }}
-    }}
+class {name}
+{{
+    private static int TestMethod2(ReadOnlySpan<char> input) => input.Length;
 
-    private static void DoWhile() {{
-        do
-        {{
-            Span<int> span = stackalloc int[1024];
-        }}
-        while (true);
+    public int TestMethod(string input)
+    {{
+        return TestMethod2(input[3..5]) + TestMethod2(input[1..^2]);
     }}
 }}
 "));
@@ -51,17 +40,17 @@ unsafe class {name} {{
 
             var compilation = CSharpCompilationHelper.CreateAsync(sources.ToArray()).GetAwaiter().GetResult();
             BaselineCompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new EmptyAnalyzer()));
-            CompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new CSharpDoNotUseStackallocInLoopsAnalyzer()));
+            CompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new UseAsSpanInsteadOfRangeIndexerAnalyzer()));
         }
 
         private static CompilationWithAnalyzers BaselineCompilationWithAnalyzers;
         private static CompilationWithAnalyzers CompilationWithAnalyzers;
 
         [Benchmark]
-        public void CA2014_DiagnosticsProduced()
+        public async Task CA1831_DiagnosticsProduced()
         {
-            var analysisResult = CompilationWithAnalyzers.GetAnalysisResultAsync(default).GetAwaiter().GetResult();
-            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.Single());
+            var analysisResult = await CompilationWithAnalyzers.GetAnalysisResultAsync(CancellationToken.None);
+            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.First());
             if (analysisResult.Analyzers.Length != 1)
             {
                 throw new InvalidOperationException($"Expected a single analyzer but found '{analysisResult.Analyzers.Length}'");
@@ -72,17 +61,17 @@ unsafe class {name} {{
                 throw new InvalidOperationException($"Expected no compilation diagnostics but found '{analysisResult.CompilationDiagnostics.Count}'");
             }
 
-            if (diagnostics.Length != 3 * Constants.Number_Of_Code_Files)
+            if (diagnostics.Length != 2 * Constants.Number_Of_Code_Files)
             {
-                throw new InvalidOperationException($"Expected '3,000' analyzer diagnostics but found '{diagnostics.Length}'");
+                throw new InvalidOperationException($"Expected '{2 * Constants.Number_Of_Code_Files:N0}' analyzer diagnostics but found '{diagnostics.Length}'");
             }
         }
 
         [Benchmark(Baseline = true)]
-        public void CA2014_Baseline()
+        public async Task CA1831_Baseline()
         {
-            var analysisResult = BaselineCompilationWithAnalyzers.GetAnalysisResultAsync(default).GetAwaiter().GetResult();
-            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.Single());
+            var analysisResult = await BaselineCompilationWithAnalyzers.GetAnalysisResultAsync(CancellationToken.None);
+            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.First());
             if (analysisResult.Analyzers.Length != 1)
             {
                 throw new InvalidOperationException($"Expected a single analyzer but found '{analysisResult.Analyzers.Length}'");
