+++ src/PerformanceTests/Tests/Enabled/CSharp_CA2014.cs
@@ -4,49 +4,66 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 using BenchmarkDotNet.Attributes;
 using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.NetCore.Analyzers.Publish;
+using Microsoft.NetCore.CSharp.Analyzers.Runtime;
 using PerformanceTests.Utilities;
 using PerfUtilities;
 
-namespace VisualBasicPerformanceTests.Enabled
+namespace CSharpPerformanceTests.Enabled
 {
-    public class VisualBasic_IL3000
+    public class CSharp_CA2014
     {
         [IterationSetup]
-        public static void CreateEnvironmentIL3000()
+        public static void CreateEnvironmentCA2014()
         {
             var sources = new List<(string name, string content)>();
             for (var i = 0; i < Constants.Number_Of_Code_Files; i++)
             {
                 var name = "TypeName" + i;
                 sources.Add((name, @$"
-Imports System
-Imports System.Reflection
+using System;
+unsafe class {name} {{
+    private static void ForLoop() {{
+        for (int i = 0; i < 10; i++)
+        {{
+            Span<byte> span = stackalloc byte[1024]|;
+        }}
+    }}
 
-Class {name}
-    Sub M()
-        Dim a = Assembly.GetExecutingAssembly()
-        Dim b = a.Location
-        Dim c = a.EscapedCodeBase
-    End Sub
-End Class"));
+    private static void WhileLoop() {{
+        while (true)
+        {{
+            Span<char> span = stackalloc char[1024];
+        }}
+    }}
+
+    private static void DoWhile() {{
+        do
+        {{
+            Span<int> span = stackalloc int[1024];
+        }}
+        while (true);
+    }}
+}}
+"));
             }
 
-            var (compilation, options) = VisualBasicCompilationHelper.CreateWithOptionsAsync(sources.ToArray(), new[] { ("build_property.PublishSingleFile", "true") }).GetAwaiter().GetResult();
-            BaselineCompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new EmptyAnalyzer()), options);
-            CompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new AvoidAssemblyLocationInSingleFile()), options);
+            var compilation = CSharpCompilationHelper.CreateAsync(sources.ToArray()).GetAwaiter().GetResult();
+            BaselineCompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new EmptyAnalyzer()));
+            CompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new CSharpDoNotUseStackallocInLoopsAnalyzer()));
         }
 
         private static CompilationWithAnalyzers BaselineCompilationWithAnalyzers;
         private static CompilationWithAnalyzers CompilationWithAnalyzers;
 
         [Benchmark]
-        public void IL3000_DiagnosticsProduced()
+        public async Task CA2014_DiagnosticsProduced()
         {
-            var analysisResult = CompilationWithAnalyzers.GetAnalysisResultAsync(default).GetAwaiter().GetResult();
-            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.Single());
+            var analysisResult = await CompilationWithAnalyzers.GetAnalysisResultAsync(CancellationToken.None);
+            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.First());
             if (analysisResult.Analyzers.Length != 1)
             {
                 throw new InvalidOperationException($"Expected a single analyzer but found '{analysisResult.Analyzers.Length}'");
@@ -57,17 +74,17 @@ End Class"));
                 throw new InvalidOperationException($"Expected no compilation diagnostics but found '{analysisResult.CompilationDiagnostics.Count}'");
             }
 
-            if (diagnostics.Length != 1 * Constants.Number_Of_Code_Files)
+            if (diagnostics.Length != 3 * Constants.Number_Of_Code_Files)
             {
-                throw new InvalidOperationException($"Expected '1,000' analyzer diagnostics but found '{diagnostics.Length}'");
+                throw new InvalidOperationException($"Expected '{3 * Constants.Number_Of_Code_Files:N0}' analyzer diagnostics but found '{diagnostics.Length}'");
             }
         }
 
         [Benchmark(Baseline = true)]
-        public void IL3000_Baseline()
+        public async Task CA2014_Baseline()
         {
-            var analysisResult = BaselineCompilationWithAnalyzers.GetAnalysisResultAsync(default).GetAwaiter().GetResult();
-            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.Single());
+            var analysisResult = await BaselineCompilationWithAnalyzers.GetAnalysisResultAsync(CancellationToken.None);
+            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.First());
             if (analysisResult.Analyzers.Length != 1)
             {
                 throw new InvalidOperationException($"Expected a single analyzer but found '{analysisResult.Analyzers.Length}'");
