+++ src/PerformanceTests/Tests/Enabled/VisualBasic_CA1416.cs
@@ -4,6 +4,8 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 using BenchmarkDotNet.Attributes;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.NetCore.Analyzers.InteropServices;
@@ -12,10 +14,10 @@ using PerfUtilities;
 
 namespace VisualBasicPerformanceTests.Enabled
 {
-    public class VisualBasic_CA1417
+    public class VisualBasic_CA1416
     {
         [IterationSetup]
-        public static void CreateEnvironmentCA1417()
+        public static void CreateEnvironmentCA1416()
         {
             var sources = new List<(string name, string content)>();
             for (var i = 0; i < Constants.Number_Of_Code_Files; i++)
@@ -23,35 +25,50 @@ namespace VisualBasicPerformanceTests.Enabled
                 var name = "TypeName" + i;
                 sources.Add((name, @$"
 Imports System
-Imports System.Runtime.InteropServices
+Imports PlatformCompatDemo.SupportedUnupported
 
-Class {name}
+Public Class {name}
 
-    <DllImport(""user32.dll"", CharSet:=CharSet.Unicode)>
-    Private Shared Sub Method1(<Out> s As String)
-    End Sub
+    Private field As B = New B()
 
-    < DllImport(""user32.dll"", CharSet:= CharSet.Unicode) >
-    Private Shared Sub Method2(s1 As String, <Out> s2 As String, <Out> s3 As String)
+    Public Sub M1()
+        field.M2()
     End Sub
 End Class
-
 "));
             }
 
-            var compilation = VisualBasicCompilationHelper.CreateAsync(sources.ToArray()).GetAwaiter().GetResult();
-            BaselineCompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new EmptyAnalyzer()));
-            CompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new DoNotUseOutAttributeStringPInvokeParametersAnalyzer()));
+            var targetTypesForTest = @"
+Imports System.Runtime.Versioning
+
+Namespace PlatformCompatDemo.SupportedUnupported
+    Public Class B
+        <SupportedOSPlatform(""Windows10.1.1.1"")>
+        Public Sub M2()
+        End Sub
+    End Class
+End Namespace
+";
+            sources.Add((nameof(targetTypesForTest), targetTypesForTest));
+            var properties = new[]
+            {
+                ("build_property.TargetFramework", "net6"),
+                ("build_property._SupportedPlatformList", "Linux,macOS"),
+            };
+
+            var (compilation, options) = VisualBasicCompilationHelper.CreateWithOptionsAsync(sources.ToArray(), properties).GetAwaiter().GetResult();
+            BaselineCompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new EmptyAnalyzer()), options);
+            CompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new PlatformCompatibilityAnalyzer()), options);
         }
 
         private static CompilationWithAnalyzers BaselineCompilationWithAnalyzers;
         private static CompilationWithAnalyzers CompilationWithAnalyzers;
 
         [Benchmark]
-        public void CA1417_DiagnosticsProduced()
+        public async Task CA1416_DiagnosticsProduced()
         {
-            var analysisResult = CompilationWithAnalyzers.GetAnalysisResultAsync(default).GetAwaiter().GetResult();
-            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.Single());
+            var analysisResult = await CompilationWithAnalyzers.GetAnalysisResultAsync(CancellationToken.None);
+            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.First());
             if (analysisResult.Analyzers.Length != 1)
             {
                 throw new InvalidOperationException($"Expected a single analyzer but found '{analysisResult.Analyzers.Length}'");
@@ -62,17 +79,17 @@ End Class
                 throw new InvalidOperationException($"Expected no compilation diagnostics but found '{analysisResult.CompilationDiagnostics.Count}'");
             }
 
-            if (diagnostics.Length != 3 * Constants.Number_Of_Code_Files)
+            if (diagnostics.Length != 1 * Constants.Number_Of_Code_Files)
             {
-                throw new InvalidOperationException($"Expected '3,000' analyzer diagnostics but found '{diagnostics.Length}'");
+                throw new InvalidOperationException($"Expected '{1 * Constants.Number_Of_Code_Files:N0}' analyzer diagnostics but found '{diagnostics.Length}'");
             }
         }
 
         [Benchmark(Baseline = true)]
-        public void CA1417_Baseline()
+        public async Task CA1416_Baseline()
         {
-            var analysisResult = BaselineCompilationWithAnalyzers.GetAnalysisResultAsync(default).GetAwaiter().GetResult();
-            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.Single());
+            var analysisResult = await BaselineCompilationWithAnalyzers.GetAnalysisResultAsync(CancellationToken.None);
+            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.First());
             if (analysisResult.Analyzers.Length != 1)
             {
                 throw new InvalidOperationException($"Expected a single analyzer but found '{analysisResult.Analyzers.Length}'");
