+++ src/PerformanceTests/Tests/Enabled/VisualBasic_CA2015.cs
@@ -4,51 +4,69 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 using BenchmarkDotNet.Attributes;
 using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.NetCore.Analyzers.InteropServices;
+using Microsoft.NetCore.Analyzers.Runtime;
 using PerformanceTests.Utilities;
 using PerfUtilities;
 
-namespace CSharpPerformanceTests.Enabled
+namespace VisualBasicPerformanceTests.Enabled
 {
-    public class CSharp_CA1417
+    public class VisualBasic_CA2015
     {
         [IterationSetup]
-        public static void CreateEnvironmentCA1417()
+        public static void CreateEnvironmentCA2015()
         {
             var sources = new List<(string name, string content)>();
             for (var i = 0; i < Constants.Number_Of_Code_Files; i++)
             {
                 var name = "TypeName" + i;
                 sources.Add((name, @$"
-using System;
-using System.Runtime.InteropServices;
+Imports System
+Imports System.Buffers
 
-class {name}
-{{
-    [DllImport(""user32.dll"", CharSet=CharSet.Unicode)]
-    private static extern void Method1([Out] string s); // Should not have [Out] string
+Class {name}(Of T)
+    Inherits MemoryManager(Of T)
 
-    [DllImport(""user32.dll"", CharSet=CharSet.Unicode)]
-    private static extern void Method2(string s1, [Out] string s2, [Out] string s3); // Should not have [Out] string
-}}
+    <Obsolete>
+    Public Overrides Function GetSpan() As Span(Of T)
+        Throw New NotImplementedException()
+    End Function
+
+    Public Overrides Function Pin(Optional elementIndex As Integer = 0) As MemoryHandle
+        Throw New NotImplementedException()
+    End Function
+
+    Public Overrides Sub Unpin()
+        Throw New NotImplementedException()
+    End Sub
+
+    Protected Overrides Sub Dispose(disposing As Boolean)
+        Throw New NotImplementedException()
+    End Sub
+
+    Protected Overrides Sub Finalize()
+    End Sub
+
+End Class
 "));
             }
 
-            var compilation = CSharpCompilationHelper.CreateAsync(sources.ToArray()).GetAwaiter().GetResult();
+            var compilation = VisualBasicCompilationHelper.CreateAsync(sources.ToArray()).GetAwaiter().GetResult();
             BaselineCompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new EmptyAnalyzer()));
-            CompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new DoNotUseOutAttributeStringPInvokeParametersAnalyzer()));
+            CompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new DoNotDefineFinalizersForTypesDerivedFromMemoryManager()));
         }
 
         private static CompilationWithAnalyzers BaselineCompilationWithAnalyzers;
         private static CompilationWithAnalyzers CompilationWithAnalyzers;
 
         [Benchmark]
-        public void CA1417_DiagnosticsProduced()
+        public async Task CA2015_DiagnosticsProduced()
         {
-            var analysisResult = CompilationWithAnalyzers.GetAnalysisResultAsync(default).GetAwaiter().GetResult();
-            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.Single());
+            var analysisResult = await CompilationWithAnalyzers.GetAnalysisResultAsync(CancellationToken.None);
+            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.First());
             if (analysisResult.Analyzers.Length != 1)
             {
                 throw new InvalidOperationException($"Expected a single analyzer but found '{analysisResult.Analyzers.Length}'");
@@ -59,17 +77,17 @@ class {name}
                 throw new InvalidOperationException($"Expected no compilation diagnostics but found '{analysisResult.CompilationDiagnostics.Count}'");
             }
 
-            if (diagnostics.Length != 3 * Constants.Number_Of_Code_Files)
+            if (diagnostics.Length != 1 * Constants.Number_Of_Code_Files)
             {
-                throw new InvalidOperationException($"Expected '3,000' analyzer diagnostics but found '{diagnostics.Length}'");
+                throw new InvalidOperationException($"Expected '{1 * Constants.Number_Of_Code_Files:N0}' analyzer diagnostics but found '{diagnostics.Length}'");
             }
         }
 
         [Benchmark(Baseline = true)]
-        public void CA1417_Baseline()
+        public async Task CA2015_Baseline()
         {
-            var analysisResult = BaselineCompilationWithAnalyzers.GetAnalysisResultAsync(default).GetAwaiter().GetResult();
-            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.Single());
+            var analysisResult = await BaselineCompilationWithAnalyzers.GetAnalysisResultAsync(CancellationToken.None);
+            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.First());
             if (analysisResult.Analyzers.Length != 1)
             {
                 throw new InvalidOperationException($"Expected a single analyzer but found '{analysisResult.Analyzers.Length}'");
