+++ src/PerformanceTests/Tests/Enabled/VisualBasic_IL3001.cs
@@ -4,69 +4,50 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 using BenchmarkDotNet.Attributes;
 using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.NetCore.Analyzers.InteropServices;
+using Microsoft.NetCore.Analyzers.Publish;
 using PerformanceTests.Utilities;
 using PerfUtilities;
 
 namespace VisualBasicPerformanceTests.Enabled
 {
-    public class VisualBasic_CA1416
+    public class VisualBasic_IL3001
     {
         [IterationSetup]
-        public static void CreateEnvironmentCA1416()
+        public static void CreateEnvironmentIL3001()
         {
             var sources = new List<(string name, string content)>();
             for (var i = 0; i < Constants.Number_Of_Code_Files; i++)
             {
                 var name = "TypeName" + i;
-                sources.Add((name, @$"
-Imports System
-Imports PlatformCompatDemo.SupportedUnupported
+                sources.Add((name, $@"Imports System
+Imports System.Reflection
 
-Public Class {name}
-
-    Private field As B = New B()
-
-    Public Sub M1()
-        field.M2()
+Class {name}
+    Sub M()
+        Dim a = Assembly.LoadFrom(""/some/path/not/in/bundle"")
+        Dim b = a.GetFile(""/some/file/path"")
+        Dim c = a.GetFiles()
     End Sub
-End Class
-"));
+End Class"));
             }
 
-            var targetTypesForTest = @"
-Imports System.Runtime.Versioning
-
-Namespace PlatformCompatDemo.SupportedUnupported
-    Public Class B
-        <SupportedOSPlatform(""Windows10.1.1.1"")>
-        Public Sub M2()
-        End Sub
-    End Class
-End Namespace
-";
-            sources.Add((nameof(targetTypesForTest), targetTypesForTest));
-            var properties = new[]
-            {
-                ("build_property.TargetFramework", "net6"),
-                ("build_property._SupportedPlatformList", "Linux,macOS"),
-            };
-
-            var (compilation, options) = VisualBasicCompilationHelper.CreateWithOptionsAsync(sources.ToArray(), properties).GetAwaiter().GetResult();
+            var (compilation, options) = VisualBasicCompilationHelper.CreateWithOptionsAsync(sources.ToArray(), new[] { ("build_property.PublishSingleFile", "true") }).GetAwaiter().GetResult();
             BaselineCompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new EmptyAnalyzer()), options);
-            CompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new PlatformCompatibilityAnalyzer()), options);
+            CompilationWithAnalyzers = compilation.WithAnalyzers(ImmutableArray.Create<DiagnosticAnalyzer>(new AvoidAssemblyLocationInSingleFile()), options);
         }
 
         private static CompilationWithAnalyzers BaselineCompilationWithAnalyzers;
         private static CompilationWithAnalyzers CompilationWithAnalyzers;
 
         [Benchmark]
-        public void CA1416_DiagnosticsProduced()
+        public async Task IL3001_DiagnosticsProduced()
         {
-            var analysisResult = CompilationWithAnalyzers.GetAnalysisResultAsync(default).GetAwaiter().GetResult();
-            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.Single());
+            var analysisResult = await CompilationWithAnalyzers.GetAnalysisResultAsync(CancellationToken.None);
+            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.First());
             if (analysisResult.Analyzers.Length != 1)
             {
                 throw new InvalidOperationException($"Expected a single analyzer but found '{analysisResult.Analyzers.Length}'");
@@ -77,17 +58,17 @@ End Namespace
                 throw new InvalidOperationException($"Expected no compilation diagnostics but found '{analysisResult.CompilationDiagnostics.Count}'");
             }
 
-            if (diagnostics.Length != 1 * Constants.Number_Of_Code_Files)
+            if (diagnostics.Length != 2 * Constants.Number_Of_Code_Files)
             {
-                throw new InvalidOperationException($"Expected '1,000' analyzer diagnostics but found '{diagnostics.Length}'");
+                throw new InvalidOperationException($"Expected '{2 * Constants.Number_Of_Code_Files:N0}' analyzer diagnostics but found '{diagnostics.Length}'");
             }
         }
 
         [Benchmark(Baseline = true)]
-        public void CA1416_Baseline()
+        public async Task IL3001_Baseline()
         {
-            var analysisResult = BaselineCompilationWithAnalyzers.GetAnalysisResultAsync(default).GetAwaiter().GetResult();
-            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.Single());
+            var analysisResult = await BaselineCompilationWithAnalyzers.GetAnalysisResultAsync(CancellationToken.None);
+            var diagnostics = analysisResult.GetAllDiagnostics(analysisResult.Analyzers.First());
             if (analysisResult.Analyzers.Length != 1)
             {
                 throw new InvalidOperationException($"Expected a single analyzer but found '{analysisResult.Analyzers.Length}'");
