--- src/PerformanceTests/VisualBasic/Program.cs
@@ -1,12 +1,0 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-using BenchmarkDotNet.Running;
-
-namespace PerformanceTests
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-            => BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
-    }
-}
